Refactored install command
