+++ synchronize-with-npm/src/index.ts
@@ -26,10 +26,12 @@ export interface ActionPayload {
 export function* run({ octokit, core, payload }: ActionPayload): Operation<void> {
   let { pkgsToPublish, pkgsToDeprecate } = listPackages();
   let confirmedPkgsToPublish = yield checkIfPublished({ pkgsToPublish });
+
   let installScript = core.getInput("INSTALL_SCRIPT") || "";
   if (confirmedPkgsToPublish.length > 0) {
-    yield install(installScript);
+    yield install({ installScript });
   }
+
   let publishedPackages: ToPublish[] = yield publishAndTag({ confirmedPkgsToPublish, octokit, payload });
   let deprecatedPackages: string[] = yield deprecatePackages({ pkgsToDeprecate });
